# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram Reels Tracking & Control Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /instagram-accounts:
    get:
      operationId: listInstagramAccounts
      tags: [instagram-accounts]
      summary: List Instagram accounts
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of Instagram accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstagramAccount"
    post:
      operationId: createInstagramAccount
      tags: [instagram-accounts]
      summary: Create Instagram account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstagramAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramAccount"
  /instagram-accounts/{id}:
    get:
      operationId: getInstagramAccount
      tags: [instagram-accounts]
      summary: Get Instagram account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Instagram account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramAccount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInstagramAccount
      tags: [instagram-accounts]
      summary: Update Instagram account (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstagramAccountBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramAccount"

  # DAILY REPORTS
  /daily-reports:
    get:
      operationId: listDailyReports
      tags: [daily-reports]
      summary: List daily reports
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of daily reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReport"
    post:
      operationId: createDailyReport
      tags: [daily-reports]
      summary: Create or get daily report for a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyReportBody"
      responses:
        "200":
          description: Report created or fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /daily-reports/{id}:
    get:
      operationId: getDailyReport
      tags: [daily-reports]
      summary: Get daily report detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily report detail with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReportDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDailyReport
      tags: [daily-reports]
      summary: Update daily report (submit, admin review)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyReportBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"

  # REPORT ITEMS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        personnelNo:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - username
        - role
        - status
        - createdAt
        - updatedAt
...
    InstagramAccount:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        instagramUsername:
          type: string
        profileUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        userName:
          type: ["string", "null"]
        userPersonnelNo:
          type: ["integer", "null"]
      required:
        - id
        - userId
        - instagramUsername
        - status
        - createdAt
        - updatedAt
    CreateInstagramAccountBody:
      type: object
      properties:
        userId:
          type: integer
        instagramUsername:
          type: string
        profileUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - userId
        - instagramUsername
    UpdateInstagramAccountBody:
      type: object
      properties:
        instagramUsername:
          type: string
        profileUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
        userId:
          type: ["integer", "null"]
    DailyReport:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        userName:
          type: ["string", "null"]
        userPersonnelNo:
          type: ["integer", "null"]
        itemCount:
          type: integer
      required:
        - id
        - userId
        - date
        - status
        - createdAt
        - updatedAt
        - itemCount
    DailyReportDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        status:
          type: string
        adminNote:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        userName:
          type: ["string", "null"]
        userPersonnelNo:
          type: ["integer", "null"]
        itemCount:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReportItemWithAccount"
      required:
        - id
        - userId
        - date
        - status
        - createdAt
        - updatedAt
        - itemCount
        - items
    CreateDailyReportBody:
      type: object
      properties:
        date:
          type: string
        userId:
          type: ["integer", "null"]
      required:
        - date
    UpdateDailyReportBody:
      type: object
      properties:
        status:
          type: string
        adminNote:
          type: ["string", "null"]
...
    ReportItemWithAccount:
      type: object
      properties:
        id:
          type: integer
        reportId:
          type: integer
        instagramAccountId:
          type: integer
        reelsUrl:
          type: string
        contentDate:
          type: string
        enteredAt:
          type: string
        createdAt:
          type: string
        instagramUsername:
          type: ["string", "null"]
      required:
        - id
        - reportId
        - instagramAccountId
        - reelsUrl
        - contentDate
        - enteredAt
        - createdAt
...