# lib/api-zod/src/generated/types/createInstagramAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Reels Tracking & Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInstagramAccountBody {
  userId: number;
  instagramUsername: string;
  /** @nullable */
  profileUrl?: string | null;
  /** @nullable */
  description?: string | null;
}