# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Reels Tracking & Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyReport {
  id: number;
  userId: number;
  date: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userPersonnelNo?: number | null;
  itemCount: number;
}