# lib/api-zod/src/generated/types/updateInstagramAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Reels Tracking & Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInstagramAccountBody {
  instagramUsername?: string;
  /** @nullable */
  profileUrl?: string | null;
  /** @nullable */
  description?: string | null;
  status?: string;
  /** @nullable */
  userId?: number | null;
}