# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Instagram Reels Tracking & Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  username: string;
  role: string;
  status: string;
  /** @nullable */
  personnelNo?: number | null;
  createdAt: string;
  updatedAt: string;
}